requirements.txt
langchain
langchain-openai
langgraph
openai
python-dotenv
pydub
fastapi
uvicorn
python-multipart
